% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.r
\name{benchML}
\alias{benchML}
\title{Run a Benchmark experiment}
\usage{
benchML(tasks, learners, outer)
}
\arguments{
\item{tasks}{List of tasks created with \code{pseq2mlr} method}

\item{learners}{List of learners created with \code{setHyperparameters} method}
}
\value{
a BenchmarkResult object generated from \code{mlr3}
}
\description{
Run a Benchmark experiment
}
