% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.r
\name{filterCFS}
\alias{filterCFS}
\title{Feature Selection - CFS}
\usage{
filterCFS(pseqMLR, cutoff)
}
\arguments{
\item{pseqMLR}{\code{pseqML} object}

\item{cutoff}{Numeric value between 0-1 of correlation}
}
\description{
Feature Selection - CFS
}
