% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.r
\name{filterFS}
\alias{filterFS}
\title{Feature Selection - FILTER}
\usage{
filterFS(pseqMLR, method)
}
\arguments{
\item{pseqMLR}{\code{pseqML} object}

\item{method}{Character vector with methods availables in \code{listFilterMethods}}
}
\description{
Feature Selection - FILTER
}
