% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.r
\name{filterMBF}
\alias{filterMBF}
\title{Feature Selection - Markov Blanket Filter}
\usage{
filterMBF(pseqMLR, threshold)

filterMBF(pseqMLR, threshold)
}
\arguments{
\item{pseqMLR}{\code{pseqML} object}

\item{threshold}{Numeric value indicating the minimum correlation (as determined by symettrical uncertainty) between each variable and the class}

\item{cutoff}{Numeric value between 0-1 of correlation}
}
\description{
Feature Selection - Markov Blanket Filter

Feature Selection - Fast Correlation Based Feature Selection
}
