% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML-Methods.r
\name{init_outer}
\alias{init_outer}
\title{Initialize a resampling method to carried out outer CV}
\usage{
init_outer(object)
}
\arguments{
\item{object}{of class \code{outer} (required)}
}
\description{
Initialize a resampling method to carried out outer CV
}
