% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML-Methods.r
\name{init_resampling}
\alias{init_resampling}
\title{Initialize a resampling class}
\usage{
init_resampling(object)
}
\arguments{
\item{object}{of class \code{resampling} (required)}
}
\description{
Initialize a resampling class
}
