% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML-Methods.r
\name{pseq2mlr}
\alias{pseq2mlr}
\title{Convert \code{phyloseq} to \code{mlr3} task}
\usage{
pseq2mlr(phyloseq, target, id)
}
\arguments{
\item{phyloseq}{Object of class \code{phyloseq} (required)}

\item{target}{Character with the name of dependent variable (required)}

\item{id}{Character vector with ID to data (required)}
}
\description{
Convert \code{phyloseq} to \code{mlr3} task
}
