% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r
\docType{class}
\name{pseqML-class}
\alias{pseqML-class}
\title{The S4 Class for storing phyloseq data to carried out ML analyisis.}
\description{
This object stores the all information of metagenomic data
in one \code{\link{phyloseq}} class. In addition, ...
}
