% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r
\docType{class}
\name{pseqMLR-class}
\alias{pseqMLR-class}
\title{The S4 class inherited from \code{pseqML} to storage also \code{TaskClassif} data}
\description{
The S4 class inherited from \code{pseqML} to storage also \code{TaskClassif} data
}
