% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.r
\name{rsmplML}
\alias{rsmplML}
\title{Carry out a resampling strategy}
\usage{
rsmplML(task, learner, outer)
}
\arguments{
\item{task}{Task created with \code{pseq2mlr} method}

\item{learner}{Learner created with \code{setHyperparameters} method}

\item{outer}{Resampling object initialized with \code{init_resampling} method}
}
\value{
a BenchmarkResult object generated from \code{mlr3}
}
\description{
Carry out a resampling strategy
}
