% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r
\docType{class}
\name{search-class}
\alias{search-class}
\title{The S4 Class for storing search strategy in CV-inner.}
\description{
This object stores all arguments to carry out search strategy in a inner CV
}
