% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS-Methods.r
\name{selectFeatures}
\alias{selectFeatures}
\title{Method to select features of a \code{pseqMLR} object}
\usage{
selectFeatures(pseqMLR, features)
}
\arguments{
\item{pseqMLR}{Object of class \code{pseqMLR}}

\item{features}{Character vector of samples to be kept}
}
\description{
Method to select features of a \code{pseqMLR} object
}
