% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML-Methods.r
\name{setHyperparameters}
\alias{setHyperparameters}
\title{Initialize a model with a set of hyperparameers}
\usage{
setHyperparameters(model, inner)
}
\arguments{
\item{object}{of class \code{inner} (required)}
}
\description{
Initialize a model with a set of hyperparameers
}
