% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.r
\name{trainML}
\alias{trainML}
\title{Training a ML algorithm}
\usage{
trainML(task, learner, predict = F, test = NULL)
}
\arguments{
\item{task}{Task created with \code{pseq2mlr} method}

\item{learner}{Learner created with \code{setHyperparameters} method}

\item{predict}{Logical value. If \code{predict = TRUE} a test tasks must be provided}

\item{test}{Task created with \code{pseq2mlr} method}
}
\value{
a Train object generated from \code{mlr3}
}
\description{
Training a ML algorithm
}
